refactor: Create a getAvailableAppHighlightAlerts function and move it in another file
