+++ src/components/AppHighlightAlert/helpers.js
@@ -1,0 +1,22 @@
+import { getBackupAppHighlightAlert } from 'components/AppHighlightAlert/BackupAppHighlightAlert'
+import { getGeolocationTrackingAppHighlightAlert } from 'components/AppHighlightAlert/GeolocationTrackingAppHighlightAlert'
+
+export const getAvailableAppHighlightAlerts = async client => {
+  const appHighlightAlerts = [
+    getBackupAppHighlightAlert(),
+    await getGeolocationTrackingAppHighlightAlert(client)
+  ]
+
+  const availableAppHighlightAlerts = appHighlightAlerts.filter(
+    status => status.available
+  )
+
+  const selectedIndex = availableAppHighlightAlerts.findIndex(
+    status => status.displayable
+  )
+  if (selectedIndex !== -1) {
+    availableAppHighlightAlerts[selectedIndex].displayed = true
+  }
+
+  return availableAppHighlightAlerts
+}
